EC-322 Moved migrations from Startup to Program.cs
