+++ Comments.Test/IntegrationTests/API/Comments/CommentTests.Insert_Multiple_Comments_And_Read_Them.expected.txt
@@ -1,12 +1,14 @@
 {
     "comments":[
         {
-            "commentId":3,
+            "commentId":39,
             "lastModifiedDate":"scrubbed by ScrubLastModifiedDate",
             "lastModifiedByUserId":"00000000-0000-0000-0000-000000000000",
             "commentText":"comment text",
+            "status":null,
+            "statusId":1,
             "commentOn":"Chapter",
-            "locationId":2,
+            "locationId":49,
             "sourceURI":"consultations://./consultation/1/document/1/chapter/introduction",
             "htmlElementID":null,
             "rangeStart":null,
@@ -16,12 +18,14 @@
             "quote":null
         },
         {
-            "commentId":1,
+            "commentId":37,
             "lastModifiedDate":"scrubbed by ScrubLastModifiedDate",
             "lastModifiedByUserId":"00000000-0000-0000-0000-000000000000",
             "commentText":"comment text",
+            "status":null,
+            "statusId":1,
             "commentOn":"Chapter",
-            "locationId":1,
+            "locationId":48,
             "sourceURI":"consultations://./consultation/1/document/1/chapter/introduction",
             "htmlElementID":null,
             "rangeStart":null,
@@ -31,12 +35,14 @@
             "quote":null
         },
         {
-            "commentId":2,
+            "commentId":38,
             "lastModifiedDate":"scrubbed by ScrubLastModifiedDate",
             "lastModifiedByUserId":"00000000-0000-0000-0000-000000000000",
             "commentText":"comment text",
+            "status":null,
+            "statusId":1,
             "commentOn":"Chapter",
-            "locationId":1,
+            "locationId":48,
             "sourceURI":"consultations://./consultation/1/document/1/chapter/introduction",
             "htmlElementID":null,
             "rangeStart":null,
