+++ Comments/Program.cs
@@ -1,5 +1,9 @@
+using System;
+using Comments.Models;
 using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.DependencyInjection;
 
 namespace Comments
 {
@@ -7,8 +11,25 @@ namespace Comments
     {
         public static void Main(string[] args)
         {
-            BuildWebHost(args).Run();
-        }
+			var host = BuildWebHost(args);
+
+	        using (var scope = host.Services.CreateScope())
+		    {
+		        var services = scope.ServiceProvider;
+		        
+			    try
+		        {
+			        var context = services.GetService<ConsultationsContext>();
+			        context.Database.Migrate();
+			    }
+		        catch (Exception e)
+		        {
+			        //TODO: Logging	
+				    throw;
+			    }
+		    }
+	        host.Run();
+		}
 
         public static IWebHost BuildWebHost(string[] args) =>
             WebHost.CreateDefaultBuilder(args)
